Fix compiler warnings in ZSTDReader.hpp
- Add missing override on seek() (Clang -Winconsistent-missing-override)
- Remove redundant > 0 on bool return from jumpTableIsInitialized
  (MSVC C4804)
- Change read() return type from int to int64_t to match ZSTDSeek_read
  and avoid truncation on large reads (MSVC C4244); update Cython
  declaration accordingly

+++ indexed_zstd/indexed_zstd.pyx
@@ -30,7 +30,7 @@ cdef extern from "ZSTDReader.hpp":
         size_t tell() except +
         size_t tellCompressed() except +
         size_t size() except +
-        int read(int, char*, size_t) except +
+        long long read(int, char*, size_t) except +
         bool blockOffsetsComplete() except +
         map[size_t, size_t] blockOffsets() except +
         map[size_t, size_t] availableBlockOffsets() except +
